Move ErrorExt impl for gimli::Error into dwarf module
It arguably makes more sense to co-locate the ErrorExt impl for each
error type that we support with the functionality that is using this
error type, as opposed to it being located in a central location, as
that leads to less conditional compilation directives being spread
throughout the code base.
To that end, this change moves the ErrorExt impl for gimli::Error into
the dwarf module. While at it, add a test for increased coverage and fix
some dead code warning popping up with Rust 1.75.
